+++ src/graphics/renderer/opengl/vertex_array.cpp
@@ -40,7 +40,7 @@ void VertexArray::add_buffer(std::shared_ptr<Buffer> buffer, const BufferLayout&
 
     size_t offset = 0;
 
-    for (unsigned int i = 0; i < layout.elements.size(); i++) {
+    for (size_t i = 0; i < layout.elements.size(); i++) {
         const VertexElement& element = layout.elements[i];
 
         switch (element.type) {
